big-AGI was renamed from 'NextJS ChatGPT' a few weeks ago
Since our UI was on the list, we added several amazing features such as
speech synthesis, visual drawing, and precise token counting, and more
improvements are on the way.

We changed the name to highlight our commitment to a user experience that
will help users with their everyday work, rather than just a bare bones
'chat' because the UI is cleaner and better looking than typical chat apps.